+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.0.2
 * Added "Waves" in *waves.js*
+* Added "Halftone" in *halftone.js"
 
 ## 3.0.1
 * Upgraded to Three.js r183